Change JavaScript source to use rawgit.com
raw.github.com does not serve the corret MIME type. This is a problem when
trying to use the bookmarklet on resources such as https://gdssocialmedia.blog.gov.uk/2014/10/13/social-media-planning-its-time-for-some-perspective/

This ensures that it is served with the correct MIME type (but it will still fail on
sites which have strict security policies).